# packages/contracts/openapi/k-admission.v1.yaml
openapi: 3.1.0
info:
  title: K-Admission University Data Plane API
  version: 1.4.0
  description: >
    Reference contract for a university-local K-Admission data plane.
    The university DB is the source of truth; central sync is not in the applicant critical path.


    1.2.0 aligns the contract with the repository implementation (spec discrepancy register
    D-7, D-16, D-17, D-19, D-20, D-22, D-24, D-26, D-32, D-34, D-35, D-37, D-38, D-39, D-40).
    Paths marked with `x-kadmission-register` were added by the implementation first.


    1.3.0 adds the common `429` response (`RateLimited`, `Retry-After`) to applicant operations (D-51).


    1.4.0 connects what the implementation had but the contract did not describe (D-55 to D-61):
    the application state machine is driven by validate / payment / finalize (READY, PAYMENT_PENDING,
    PAID are now real states and editing stops at PAYMENT_PENDING); one live payment per application
    (`200` resumes the open checkout, `409 PAYMENT_IN_PROGRESS` refuses a second payment); the form
    schema carries `profileFields` and `documents` so the client draws everything from config; the
    receipt is audited; server clock offset against the DB is reported (§A9); the applicant-facing
    common profile API on the central host; the active config includes its body and config drafts
    are linted (§A5); universities send the §04 heartbeat and the dashboard shows which university
...


    Common rules:
...
servers:
...
    variables:
...
paths:
...
  /api/v1/applications:
    post:
      tags: [Applications]
      operationId: createApplication
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateApplicationRequest' }
      responses:
        '201':
          description: Application created
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '409':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/applications/{applicationId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      tags: [Applications]
      operationId: getApplication
      responses:
        '200':
          description: Application
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '404':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
    patch:
      tags: [Applications]
      operationId: updateApplication
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/UpdateApplicationRequest' }
      responses:
        '200':
          description: Updated
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '412':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/applications/{applicationId}/validate:
    post:
      tags: [Applications]
      operationId: validateApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationResult' }
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /admin/v1/evidence/applications/{applicationId}:
    get:
      tags: [Admin]
      operationId: getEvidencePackage
      description: >
        Viewing is itself audited (ADMIN_VIEWED_PII) with the viewer and the reason.
        Contains no name, resident number, contact, address, application body or file originals.
      security: [{ oidc: [auditor] }]
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - in: query
          name: reason
          required: true
          x-kadmission-register: D-24
          description: Purpose of access. Empty → 400.
          schema: { type: string, minLength: 1 }
      responses:
        '200':
          description: Evidence package
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EvidencePackage' }
        '400':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://id.example.kr/.well-known/openid-configuration
    mutualTLS:
      type: mutualTLS
  parameters:
    ApplicationId:
      in: path
      name: applicationId
      required: true
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 200 }
    IfMatch:
      in: header
      name: If-Match
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: >
        Too many requests for this applicant (code RATE_LIMITED, D-51, ADR-0007). Limits are per
        authenticated applicant and request class, never per IP, so many applicants behind one school NAT
        are not blocked together. The request was not processed; retry after `Retry-After` seconds with the
        same Idempotency-Key. A normal submit (finalize) is never limited for a low-risk session.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    CreateApplicationRequest:
      type: object
      additionalProperties: false
      required: [cycleId, admissionTypeId, departmentId]
      properties:
        cycleId: { type: string, format: uuid }
        admissionTypeId: { type: string, format: uuid }
        departmentId: { type: string, format: uuid }
    UpdateApplicationRequest:
      type: object
      properties:
        admissionTypeId: { type: string, format: uuid }
        departmentId: { type: string, format: uuid }
        fields: { type: object, additionalProperties: true }
    Application:
      type: object
      required: [id, cycleId, status, version, serverTime]
      properties:
        id: { type: string, format: uuid }
        cycleId: { type: string, format: uuid }
        admissionTypeId: { type: string, format: uuid }
        departmentId: { type: string, format: uuid }
        status:
          type: string
          description: >
            DRAFT → READY (validate passed) → PAYMENT_PENDING (payment intent; no more edits) →
            PAID (payment confirmed) → FINALIZED (finalize transaction). A save moves READY back to
            DRAFT; a failed payment moves PAYMENT_PENDING back to READY. FINALIZING exists only inside
            the finalize transaction and is not stored (D-55). EXPIRED is reserved.
          enum: [DRAFT, READY, PAYMENT_PENDING, PAID, FINALIZING, FINALIZED, CANCELLED, EXPIRED]
        version: { type: integer, minimum: 1 }
        lastSavedAt: { type: [string, 'null'], format: date-time }
        fields: { type: object, additionalProperties: true }
        serverTime: { type: string, format: date-time }
    ValidationIssue:
      type: object
      required: [code, path, message]
      properties:
        code: { type: string }
        path: { type: string }
        message: { type: string }
    ValidationResult:
      type: object
      required: [valid, issues]
      properties:
        valid: { type: boolean }
        issues:
          type: array
          items: { $ref: '#/components/schemas/ValidationIssue' }
...
    EvidencePackage:
      type: object
      description: >
        Reconstruction of one application's submission process. `evidenceHash` excludes
        `generatedAt`, so two extractions of the same content hash identically.
      required: [applicationId, generatedAt, evidenceHash, application, submission, deadlinePolicy,
        configVersion, payments, documents, consents, timeline, chainVerification]
      properties:
        applicationId: { type: string, format: uuid }
        generatedAt: { type: string, format: date-time }
        evidenceHash: { type: string }
        application:
          type: object
          required: [status, version, cycleId, admissionTypeCode, departmentCode]
          properties:
            status: { type: string }
            version: { type: string }
            cycleId: { type: string, format: uuid }
            admissionTypeCode: { type: string }
            departmentCode: { type: string }
        submission:
          type: [object, 'null']
          required: [applicationNumber, requestedAt, paymentVerifiedAt, finalizedAt,
            deadlinePolicyVersion, configVersion, serverClockOffsetMs, evidenceHash]
          properties:
            applicationNumber: { type: string }
            requestedAt: { type: string, format: date-time }
            paymentVerifiedAt: { type: string, format: date-time }
            finalizedAt: { type: string, format: date-time }
            deadlinePolicyVersion: { type: string }
            configVersion: { type: string }
            serverClockOffsetMs: { type: integer }
            evidenceHash: { type: string }
        deadlinePolicy:
          type: [object, 'null']
          description: The policy actually applied at submission time.
          required: [version, mode, deadlineAt, approvedBy, activatedAt, policyHash, signedActivation]
          properties:
            version: { type: string }
            mode: { type: string }
            deadlineAt: { type: string, format: date-time }
            approvedBy:
              type: array
              items: { type: string }
            activatedAt: { type: [string, 'null'], format: date-time }
            policyHash: { type: string }
            signedActivation:
              type: [object, 'null']
              description: null when the policy was applied before signed records existed.
              required: [activationId, kind, effectiveAt, operatorId, decisionRef, keyId, signature, matchesPolicy]
              properties:
                activationId: { type: string, format: uuid }
                kind: { type: string, enum: [ACTIVATE, EXTEND, ROLLBACK] }
                effectiveAt: { type: string, format: date-time }
                operatorId: { type: string }
                decisionRef: { type: [string, 'null'] }
                keyId: { type: string }
                signature: { $ref: '#/components/schemas/SignatureStatus' }
                matchesPolicy: { type: boolean }
        configVersion:
          type: [object, 'null']
          required: [version, configHash, activatedAt]
          properties:
            version: { type: string }
            configHash: { type: string }
            activatedAt: { type: [string, 'null'], format: date-time }
        payments:
          type: array
          items:
            type: object
            required: [provider, status, amount, providerApprovedAt, verifiedAt, events]
            properties:
              provider: { type: string }
              status: { type: string }
              amount: { type: number }
              providerApprovedAt: { type: [string, 'null'], format: date-time }
              verifiedAt: { type: [string, 'null'], format: date-time }
              events:
                type: array
                items:
                  type: object
                  required: [eventType, occurredAt, payloadHash]
                  properties:
                    eventType: { type: string }
                    occurredAt: { type: [string, 'null'], format: date-time }
                    payloadHash: { type: string }
        documents:
          type: array
          items:
            type: object
            required: [documentType, status, sha256, scans]
            properties:
              documentType: { type: string }
              status: { type: string }
              sha256: { type: string }
              scans:
                type: array
                items:
                  type: object
                  required: [scanner, result, scannedAt]
                  properties:
                    scanner: { type: string }
                    engineVersion: { type: [string, 'null'] }
                    result: { type: string }
                    scannedAt: { type: [string, 'null'], format: date-time }
        consents:
          type: array
          items:
            type: object
            required: [consentCode, policyVersion, grantedAt]
            properties:
              consentCode: { type: string }
              policyVersion: { type: string }
              grantedAt: { type: string, format: date-time }
        timeline:
          type: array
          items:
            type: object
            required: [at, action, result, actorType, actorId, traceId, configVersion]
            additionalProperties: true
            properties:
              at: { type: string, format: date-time }
              action: { type: string }
              result: { type: string }
              actorType: { type: string }
              actorId: { type: [string, 'null'], description: Pseudonymous; never the original identity }
              traceId: { type: [string, 'null'] }
              configVersion: { type: [string, 'null'] }
        chainVerification: { $ref: '#/components/schemas/ChainVerification' }
    ChainVerification:
      type: object
      required: [valid, checked]
      properties:
        valid: { type: boolean }
        checked: { type: integer, minimum: 0 }
        brokenAt: { type: string, description: First event id where the hash chain breaks }
...
    SignatureStatus:
      type: string
      description: Re-verified on every read; not a stored value.
      enum: [VALID, INVALID, UNKNOWN_KEY]
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        traceId: { type: string }
        serverTime: { type: string, format: date-time }
security:
  - oidc: []